Replace PDF/docx API documentation with Markdown.
This is a proposal to move to a more friendly format so that documentation
fixes/clarification can be more easily suggested by the community.

This Markdown file contains all the content of the original docx/pdf
documents, while remaining approximately as well-formatted. Markdown has a
couple key advantages for documentation:
* Better diff behavior. docx/pdf file formats are binary, which makes diffs hard to review. Markdown is in plain text with a very readable syntax, making for sane diffs.
* Reduced file size. This markdown file is roughly half the size of the docx and 1/20th the size of the PDF.
* Easier to edit: Modifications to docx and the PDF require either Microsoft Word and a PDF editor/exporter. Modifications to markdown only require a text editor.